+++ packages/react/src/theme/foundations/index.ts
@@ -1,2 +1,4 @@
 export * from "./colors";
 export * from "./fonts";
+export * from "./shadows";
+export * from "./radii";